Buttons' should only know their parent's surface and their style should be totally configurable.
Also, they should be position regarding their parent surface
and not the world space. This implies that, to know if anything
collided with them, they should also know their parents' transform

Furthermore, to make the button fully dependant on the surface, it's
now blitted on the parent surface and not on the display screen, this
was a bad practice. Now any changes applied to the surface (changing
its alpha channel for example) will apply also to the buttons.